requirements.txt
google-ads==22.0.0
python-dotenv